calculate number of shares for equity from funds (#12)
based on https://github.com/RMI-PACTA/pacta.portfolio.analysis/pull/239

> The issue was raised that the ownership weight for equity holdings
that are held indirectly through funds always have zero ownership
weight. As far as I can tell / know, this is unintentional, and it
happens simply because we do not calculate the number of shares for
holdings from funds after they are expanded into their component
holdings. This PR fixes that.
> 
> Setting to draft mode until we collectively agree internally what we
want to do. Either way (we should or should not include indirect equity
in the ownership weight allocation), something needs to be fixed in this
repo once the decision is made... either we allocate with ownership
weight for indirect equity properly (not all 0s), or we explicitly avoid
allocating with ownership weight for indirect equity.

---------

Co-authored-by: Jackson Hoffart <jackson.hoffart@gmail.com>